% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formathobo.R
\name{trim_data}
\alias{trim_data}
\title{Trim Out-of-Water Data}
\usage{
trim_data(df, pressure_jump_thresh = 0.05, trim_start = TRUE, trim_end = TRUE)
}
\description{
Internal helper used by format_ischua().
}
\keyword{internal}
