% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formathobo.R
\name{formatischua}
\alias{formatischua}
\title{Format and Trim ISCHUA Logger Files}
\usage{
formatischua(
  file_path,
  dataset_name,
  trim_start = TRUE,
  trim_end = TRUE,
  pressure_jump_thresh = 0.05
)
}
\arguments{
\item{file_path}{Path to the CSV file.}

\item{dataset_name}{Name assigned to the dataset.}

\item{trim_start}{Logical; trim start of deployment?}

\item{trim_end}{Logical; trim end of deployment?}

\item{pressure_jump_thresh}{Numeric threshold for detecting pressure transitions.}
}
\value{
A cleaned and trimmed data frame.
}
\description{
Reads a HOBO logger CSV file, standardizes column names,
determines sampling interval, and trims out-of-water
periods at the start and end of deployment.
}
